refactor: (#26) 스피너 재사용을 위한 컴포넌트로 작성
