+++ src/pages/WebtoonSearch.tsx
@@ -6,7 +6,7 @@ import { getWeekdayLabel } from '@/utils/weekday';
 import { SearchHeader } from '@/components/SearchHeader';
 import { SearchedWebtoonItem } from '@/components/SearchedWebtoonItem';
 
-import Spinner from '@/assets/spinner.svg';
+import Spinner from '@/components/Spinner';
 
 export default function WebtoonSearch() {
   const [searchParams] = useSearchParams();
@@ -25,9 +25,7 @@ export default function WebtoonSearch() {
         <SearchHeader keyword={keyword} totalCount={searchedWebtoons.length} />
         <div className="mt-[20px]">
           {isLoading ? (
-            <div className="flex justify-center items-center py-10">
-              <img src={Spinner} alt="로딩 중" className="animate-spin h-8 w-8" />
-            </div>
+            <Spinner message="검색 결과를 불러오는 중입니다..." />
           ) : (
             <>
               {error ? (
